Incorporating `pandera` DataFrame validators
