+++ echopop/core.py
@@ -7,71 +7,6 @@ from datetime import datetime
 import numpy as np
 import pandas as pd
 
-# Required configuration initialization YAML structure
-CONFIG_INIT_MODEL = {
-    "stratified_survey_mean_parameters": {
-        "strata_transect_proportion": float,
-        "num_replicates": int,
-        "mesh_transects_per_latitude": int,
-    },
-    "bio_hake_len_bin": [float],
-    "bio_hake_age_bin": [float],
-    "TS_length_regression_parameters": {
-        "ANY": {
-            "number_code": int,
-            "TS_L_slope": float,
-            "TS_L_intercept": float,
-            "length_units": str,
-        }
-    },
-    "geospatial": {"init": str},
-    "kriging_parameters": {
-        "A0": float,
-        "longitude_reference": float,
-        "longitude_offset": float,
-        "latitude_offset": float,
-    },
-}
-
-# Required data configuration YAML structure
-CONFIG_DATA_MODEL = {
-    "survey_year": int,
-    "species": {"number_code": int},
-    "CAN_haul_offset": int,
-    "data_root_dir": str,
-    "biological": {
-        "length": {
-            "US": {"filename": str, "sheetname": str},
-            "CAN": {"filename": str, "sheetname": str},
-        },
-        "specimen": {
-            "US": {"filename": str, "sheetname": str},
-            "CAN": {"filename": str, "sheetname": str},
-        },
-        "catch": {
-            "US": {"filename": str, "sheetname": str},
-            "CAN": {"filename": str, "sheetname": str},
-        },
-        "haul_to_transect": {
-            "US": {"filename": str, "sheetname": str},
-            "CAN": {"filename": str, "sheetname": str},
-        },
-    },
-    "stratification": {
-        "strata": {"filename": str, "sheetname": str},
-        "geo_strata": {"filename": str, "sheetname": [str]},
-    },
-    "NASC": {
-        "no_age1": {"filename": str, "sheetname": str},
-        "all_ages": {"filename": str, "sheetname": str},
-    },
-    "kriging": {
-        "mesh": {"filename": str, "sheetname": str},
-        "isobath_200m": {"filename": str, "sheetname": str},
-        "vario_krig_para": {"filename": str, "sheetname": str},
-    },
-}
-
 # Name configuration dictionary
 NAME_CONFIG = {
     "Age": "age",
@@ -101,152 +36,6 @@ NAME_CONFIG = {
     "wt": "fraction_hake",
     "Weight": "weight",
     "Weight_In-Haul": "haul_weight",
-}
-
-# `Survey` object data structure
-CONFIG_MAP = {
-    "biological": {
-        # BIOLOGICAL DATASET -- LENGTH
-        "length": {
-            "haul_num": int,
-            "species_id": int,
-            "sex": np.uint8,
-            "length": np.float64,
-            "length_count": int,
-        },
-        # BIOLOGICAL DATASET -- CATCH
-        "catch": {
-            "haul_num": int,
-            "species_id": int,
-            "haul_weight": np.float64,
-        },
-        # BIOLOGICAL DATASET -- SPECIMEN
-        "specimen": {
-            "haul_num": int,
-            "species_id": int,
-            "sex": np.int8,
-            "length": np.float64,
-            "weight": np.float64,
-            "age": np.float64,
-        },
-        # BIOLOGICAL DATASET -- HAUL TO TRANSECT
-        "haul_to_transect": {
-            "haul_num": int,
-            "transect_num": int,
-        },
-    },
-    "stratification": {
-        # STRATIFICATION DATASET -- STRATA
-        "strata": {
-            "stratum_num": int,
-            "haul_num": int,
-            "fraction_hake": np.float64,
-        },
-        # STRATIFICATION DATASET -- GEOSTRATA
-        "geo_strata": {
-            "stratum_num": int,
-            "northlimit_latitude": np.float64,
-            "haul_start": int,
-            "haul_end": int,
-        },
-        "inpfc_strata": {
-            "stratum_num": int,
-            "northlimit_latitude": np.float64,
-            "haul_start": int,
-            "haul_end": int,
-        },
-    },
-    "NASC": {
-        # ACOUSTIC DATASET -- NASC EXCLUDING AGE-1
-        "no_age1": {
-            "transect_num": int,
-            "vessel_log_start": np.float64,
-            "vessel_log_end": np.float64,
-            "latitude": np.float64,
-            "longitude": np.float64,
-            "stratum_num": int,
-            "transect_spacing": np.float64,
-            "NASC": np.float64,
-            "haul_num": int,
-        },
-        # ACOUSTIC DATASET -- NASC EXCLUDING AGE-1
-        "all_ages": {
-            "transect_num": int,
-            "vessel_log_start": np.float64,
-            "vessel_log_end": np.float64,
-            "latitude": np.float64,
-            "longitude": np.float64,
-            "stratum_num": int,
-            "transect_spacing": np.float64,
-            "NASC": np.float64,
-            "haul_num": int,
-        },
-    },
-    "kriging": {
-        # STATISTICAL DATASET -- KRIGING MESH
-        "mesh": {
-            "centroid_latitude": np.float64,
-            "centroid_longitude": np.float64,
-            "fraction_cell_in_polygon": np.float64,
-        },
-        # STATISTICAL DATASET -- 200 M ISOBATH
-        "isobath_200m": {
-            "latitude": np.float64,
-            "longitude": np.float64,
-        },
-        # STATISTICAL DATASET -- KRIGING VARIOGRAM PARAMETERIZATION
-        "vario_krig_para": {
-            "dataprep.y_offset": np.float64,
-            "vario.range": np.float64,
-            "vario.res": np.float64,
-            "vario.vario": np.float64,
-            "vario.corr": np.float64,
-            "vario.nugt": np.float64,
-            "vario.sill": np.float64,
-            "vario.lscl": np.float64,
-            "vario.powr": np.float64,
-            "vario.hole": np.float64,
-            "vario.model": np.float64,
-            "vario.ytox_ratio": np.float64,
-            "vario.ztox_ratio": np.float64,
-            "vario.dim": np.float64,
-            "krig.x_res": np.float64,
-            "krig.y_res": np.float64,
-            "krig.proc_opt": np.float64,
-            "krig.load_para": np.float64,
-            "krig.vario_para": np.float64,
-            "krig.krig_para": np.float64,
-            "krig.both_para": np.float64,
-            "krig.load_griddata_file": np.float64,
-            "krig.xmin0": np.float64,
-            "krig.xmax0": np.float64,
-            "krig.dx0": np.float64,
-            "krig.ymin0": np.float64,
-            "krig.ymax0": np.float64,
-            "krig.dy0": np.float64,
-            "krig.nx": np.float64,
-            "krig.ny": np.float64,
-            "krig.nz": np.float64,
-            "krig.dx": np.float64,
-            "krig.dy": np.float64,
-            "krig.dz": np.float64,
-            "krig.model": np.float64,
-            "krig.scheme": np.float64,
-            "krig.blk_nx": np.float64,
-            "krig.blk_ny": np.float64,
-            "krig.blk_nz": np.float64,
-            "krig.srad": np.float64,
-            "krig.kmin": int,
-            "krig.kmax": int,
-            "krig.elim": np.float64,
-            "krig.eps": np.float64,
-            "krig.ratio": np.float64,
-            "krig.xmin": np.float64,
-            "krig.xmax": np.float64,
-            "krig.ymin": np.float64,
-            "krig.ymax": np.float64,
-        },
-    },
 }
 
 ECHOVIEW_EXPORT_MAP = {
